add wrapper types to API
