+++ src/Wrap/PyList.jl
@@ -1,15 +1,3 @@
-"""
-    PyList{T=Py}([x])
-
-Wraps the Python list `x` (or anything satisfying the sequence interface) as an `AbstractVector{T}`.
-
-If `x` is not a Python object, it is converted to one using `pylist`.
-"""
-struct PyList{T} <: AbstractVector{T}
-    py::Py
-    PyList{T}(x = pylist()) where {T} = new{T}(ispy(x) ? Py(x) : pylist(x))
-end
-export PyList
 
 PyList(x = pylist()) = PyList{Py}(x)
