+++ src/Wrap/PyTable.jl
@@ -1,12 +1,3 @@
-"""
-    PyTable(x)
-
-Wrap `x` as a Tables.jl-compatible table.
-
-`PyTable` is an abstract type. See [`PyPandasDataFrame`](@ref) for a concrete example.
-"""
-abstract type PyTable end
-export PyTable
 
 PyTable(x) = pyconvert(PyTable, x)
